+++ R/utils-workbook.R
@@ -1,11 +1,10 @@
-
 # Validate ----------------------------------------------------------------
 
 
 .stop_bad_input <- function(wb, content, table_name = NULL) {
 
-  if (!inherits(wb, "Workbook")) {
-    stop("'wb' must be a Workbook-class object.")
+  if (!inherits(wb, "wbWorkbook")) {
+    stop("'wb' must be an openxlsx2 wbWorkbook-class object.")
   }
 
   if (!is.null(table_name) &&
@@ -19,7 +18,6 @@
 
 
 # Detect meta elements ----------------------------------------------------
-
 
 .has_blanks_message <- function(content, tab_title) {
 
@@ -205,24 +203,24 @@
 
   if (sheet_type %in% c("cover", "contents", "notes")) {
 
-    openxlsx::writeData(
-      wb = wb,
+    wb$add_data(
       sheet = tab_title,
       x = sheet_title,
-      startCol = 1,
-      startRow = 1
+      start_col = 1,
+      start_row = 1,
+      na.strings = ""
     )
 
   }
 
   if (sheet_type == "tables") {
 
-    openxlsx::writeData(
-      wb = wb,
+    wb$add_data(
       sheet = tab_title,
       x = sheet_title,
-      startCol = 1,
-      startRow = 1
+      start_col = 1,
+      start_row = 1,
+      na.strings = ""
     )
 
   }
@@ -255,12 +253,12 @@
     ifelse(table_count == "one", "table.", "tables.")
   )
 
-  openxlsx::writeData(
-    wb = wb,
+  wb$add_data(
     sheet = tab_title,
     x = text,
-    startCol = 1,
-    startRow = 2  # table count will always be the second row
+    start_col = 1,
+    start_row = 2,  # table count will always be the second row,
+    na.strings = ""
   )
 
   wb
@@ -276,12 +274,12 @@
     text <-
       "This table contains notes, which can be found in the Notes worksheet."
 
-    openxlsx::writeData(
-      wb = wb,
+    wb$add_data(
       sheet = tab_title,
       x = text,
-      startCol = 1,
-      startRow = 3  # notes will always go in row 3 if they exist
+      start_col = 1,
+      start_row = 3,  # notes will always go in row 3 if they exist
+      na.strings = ""
     )
 
   }
@@ -300,12 +298,12 @@
     has_notes <- .has_notes(content, tab_title)
     start_row <- .get_start_row_blanks_message(has_notes)
 
-    openxlsx::writeData(
-      wb = wb,
+    wb$add_data(
       sheet = tab_title,
       x = blanks_text,
-      startCol = 1,
-      startRow = start_row
+      start_col = 1,
+      start_row = start_row,
+      na.strings = ""
     )
 
   }
@@ -331,13 +329,26 @@
 
     for (i in seq_along(custom_rows_text)) {
 
-      openxlsx::writeData(
-        wb = wb,
-        sheet = tab_title,
-        x = custom_rows_text[[i]],
-        startCol = 1,
-        startRow = start_row + (i - 1)
-      )
+      has_hyperlink <- class(custom_rows_text[[i]]) == "hyperlink"
+
+      if (has_hyperlink) {
+
+        wb$add_formula(
+          sheet = tab_title,
+          x = create_hyperlink(text = names(custom_rows_text[[i]]),
+                               file = custom_rows_text[[i]]),
+          dims = wb_dims(cols = 1, rows = start_row + (i - 1))
+        )
+      }
+
+      if (!has_hyperlink) {
+        wb$add_data(
+          sheet = tab_title,
+          x = custom_rows_text[[i]],
+          start_row = start_row + (i - 1),
+          na.strings = ""
+        )
+      }
 
     }
 
@@ -365,14 +376,26 @@
       .has_custom_rows(content, tab_title)
     )
 
-    openxlsx::writeData(
-      wb = wb,
-      sheet = tab_title,
-      x = source_text,
-      startCol = 1,
-      startRow = start_row
-    )
+    has_hyperlink <- class(source_text) == "hyperlink"
+    if (has_hyperlink) {
 
+      wb$add_formula(
+        sheet = tab_title,
+        x = create_hyperlink(text = names(source_text)[[1]],
+                             file = source_text),
+        dims = wb_dims(cols = 1, rows = start_row)
+      )
+
+    } else {
+      wb$add_data(
+        sheet = tab_title,
+        x = source_text,
+        start_col = 1,
+        start_row = start_row,
+        na.strings = ""
+      )
+
+    }
   }
 
   wb
@@ -394,16 +417,16 @@
     .has_source(content, tab_title)
   )
 
-  openxlsx::writeDataTable(
-    wb = wb,
+  wb$add_data_table(
     sheet = tab_title,
     x = table,
-    tableName = table_name,
-    startCol = 1,
-    startRow = start_row,
-    tableStyle = "none",
-    withFilter = FALSE,
-    bandedRows = FALSE
+    table_name = table_name,
+    start_col = 1,
+    start_row = start_row,
+    table_style = "none",
+    with_filter = FALSE,
+    banded_rows = FALSE,
+    na.strings = ""
   )
 
   wb
@@ -432,23 +455,24 @@
 
   for (i in seq_along(table_with_links)) {
 
-    has_hyperlink <- .detect_hyperlink(table_with_links[[i]])
+    has_hyperlink <- class(table_with_links[[i]]) == "hyperlink"
 
     if (has_hyperlink) {
-      openxlsx::writeFormula(
-        wb = wb,
+
+      wb$add_formula(
         sheet = tab_title,
-        x = table_with_links[[i]],
-        startRow = i + 1
+        x = create_hyperlink(text = names(table_with_links[[i]]),
+                             file = table_with_links[[i]]),
+        dims = wb_dims(cols = 1, rows = i + 1)
       )
     }
 
     if (!has_hyperlink) {
-      openxlsx::writeData(
-        wb = wb,
+      wb$add_data(
         sheet = tab_title,
         x = table_with_links[[i]],
-        startRow = i + 1
+        start_row = i + 1,
+        na.strings = ""
       )
     }
 
@@ -540,7 +564,7 @@
   .stop_bad_input(wb, content)
 
   for (i in unique(content$tab_title)) {
-    openxlsx::addWorksheet(wb, i)
+    wb$add_worksheet(i)
   }
 
   wb
@@ -557,14 +581,15 @@
   .insert_title(wb, content, tab_title)
   .insert_cover_table(wb, content, table_name)  # rather than .insert_table
 
-  styles <- .style_create()
-  .style_workbook(wb)
-  .style_sheet_title(wb, tab_title, styles)
-  .style_cover(wb, content, styles)  # TODO: needs special handling if list provided
+  styles <- .style_paragraph()
+  fonts <- .style_font()
+  .style_sheet_title(wb, tab_title, styles, fonts)
+  .style_cover(wb, content, styles, fonts)  # TODO: needs special handling if list provided
 
   wb
 
 }
+
 
 
 .add_contents <- function(wb, content) {
@@ -579,10 +604,10 @@
   .insert_custom_rows(wb, content, tab_title)
   .insert_table(wb, content, table_name)
 
-  styles <- .style_create()
-  .style_workbook(wb)
-  .style_sheet_title(wb, tab_title, styles)
-  .style_table(wb, content, table_name, styles)
+  styles <- .style_paragraph()
+  fonts <- .style_font()
+  .style_sheet_title(wb, tab_title, styles, fonts)
+  .style_table(wb, content, table_name, styles, fonts)
   .style_contents(wb, content, styles)
 
   wb
@@ -602,10 +627,10 @@
   .insert_custom_rows(wb, content, tab_title)
   .insert_table(wb, content, table_name)
 
-  styles <- .style_create()
-  .style_workbook(wb)
-  .style_sheet_title(wb, tab_title, styles)
-  .style_table(wb, content, table_name, styles)
+  styles <- .style_paragraph()
+  fonts <- .style_font()
+  .style_sheet_title(wb, tab_title, styles, fonts)
+  .style_table(wb, content, table_name, styles, fonts)
   .style_notes(wb, content, styles)
 
   wb
@@ -626,10 +651,10 @@
   .insert_custom_rows(wb, content, tab_title)
   .insert_table(wb, content, table_name)
 
-  styles <- .style_create()
-  .style_workbook(wb)
-  .style_sheet_title(wb, tab_title, styles)
-  .style_table(wb, content, table_name, styles)
+  styles <- .style_paragraph()
+  fonts <- .style_font()
+  .style_sheet_title(wb, tab_title, styles, fonts)
+  .style_table(wb, content, table_name, styles, fonts)
 
   wb